Multi search for Topic table page
