+++ apinf_packages/mqtt_dashboard/client/topics/table/table.less
@@ -1,5 +1,5 @@
 .table-bordered > tbody > tr > td.no-right-border {
-  border-right-color: white;
+  border-right-color: transparent;
 }
 
 #reload-table {
